Refactored how the relationship between instances and library is given.
Now you need to pass a library reference to the instance, and it will make use of this library to set up its shots.

Because this shift, the user is also able to directly use the library to build whatever they want with it.

And with this change happening, it makes much more sense now to rename instance to show_manager, since they do not represent an instance of the api, but an object that sole manages shows